refactor: move persistence implementations to data_access package and created new use case files for Angela's user story
