+++ src/main/java/use_case/generate_charts/GenerateChartsUseCase.java
@@ -1,0 +1,4 @@
+package use_case.generate_charts;
+
+public class GenerateChartsUseCase {
+}
