+++ src/main/java/use_case/view_analytics/ViewAnalyticsUseCase.java
@@ -1,0 +1,4 @@
+package use_case.view_analytics;
+
+public class ViewAnalyticsUseCase {
+}
